return to the last code

+++ src/main/java/com/example/demo/service/MemberService.java
@@ -5,10 +5,10 @@ import com.example.demo.entity.Member;
 import java.util.List;
 
 public interface MemberService {
-    Member getMemberById(Integer id);
+    Member getMemberById(Long id);
     List<Member> getAllMembers();
     List<Member> findByFirstNameOrMembershipNumberOrFamilyName(String searchTerm);
     Member addMember(Member member);
-    Member updateMember(Member member, Integer id);
-    void deleteMember(Integer id);
+    Member updateMember(Member member, Long id);
+    void deleteMember(Long id);
 }
